Gather all type constructors in "types/constructors.ts" file
