+++ src/parse.ts
@@ -4,7 +4,8 @@ import { verifyPrimitive } from './verify-primitive'
 
 import type { InvalidSubject, ErrorPath } from './error'
 import type { EitherError } from './utils/fp'
-import type { Schema, Con_Schema_SubjT } from './types/compounds'
+import type { Schema } from './types/compounds'
+import type { Con_Schema_SubjT } from './types/constructors'
 
 export function parse<T extends Schema>(
   schema: T,
