+++ src/validate.ts
@@ -3,8 +3,9 @@ import { ERROR_CODE } from './error'
 import { verifyPrimitive } from './verify-primitive'
 
 import type { EitherError } from './utils/fp'
-import type { Schema, Con_Schema_SubjT } from './types/compounds'
+import type { Schema } from './types/compounds'
 import type { InvalidSubject, ErrorPath } from './error'
+import type { Con_Schema_SubjT } from './types/constructors'
 
 export function validate<T extends Schema>(
   schema: T,
